# CB-13610 Optimize time monitor